Change default install advice b/c Rake
So people get the upcoming Rake tasks for free, let's recommend require.
This will load the railtie automatically with the Rails rake require
step.

To offset this, let's limit group to development/test. We want to be in
test so that if someone runs `rake standard test` together, the test 
task won't set RAILS_ENV to test and thereby whack standard out of 
existence