Reverse the hashrocket rule in favor of 1.9 :'s
This was a hill I was hoping to die on, but Matz's well-timed Keep Ruby
Weird talk revealed that in Ruby 3, it is very likely that hashrocket 
syntax will _not_ be able to be used to invoke methods with keyword 
arguments, which would break compatibility of, like, every single Rails
method. 

Rather than implement a style guide that would lead to a lot of future
inconsistency of when to use `:`'s and `=>`'s based on an implementation
detail of a method's signature (kwargs or an options hash)—especially
when the decision to use `=>` was justified on the grounds that they are
more consistent—seems an obviously bad idea.

So, I am defeated. 1.9 syntax wins!