Allow Standard to integrate with AtomLinter
This allows Standard to integrate with AtomLinter by adding a few minor
tweaks to the CLI and config file.

- The CLI now supports `-v` and `--version` to print out the RuboCop
  version that Standard is using under the hood. This is because
  AtomLinter requests to know the RuboCop version in order to alter the
  CLI flags it passes when linting.
- The config file now supports `silence_cta`. Setting this to `true`
  will silence the CTA message that is printed after the warning output.
  This is because AtomLinter wants to consume the RuboCop offenses as
  JSON, but having the CTA message printed at the end makes the contents
  of stdout invalid JSON. The default of this option is `false` so it
  should not impact the default behavior of the tool.

I believe these changes should enable Standard to work with other linter
tools that are compatible with RuboCop besides AtomLinter. However, I am
not a user of any other tools, so I can't vouch for compatibility one
way or the other.