Disable Style/AccessModifierDeclarations
Regardless of the merits (I think most Rubyists use the "group" style
regardless, and it's exceedingly rare to see individual methods 
habitually labeled public/private/protected) of this rule, the 
unintended consequence of enforcing this is that it is very hard to take
an existing method and privatize it by passing a symbol (e.g. 
`private :existing_method_i_inherited`). As @will pointed out, this
is enough of a pain that I don't think enforcing group-styling is worth
the cost.