Breaks down the creation of RuboCop config store
This class was getting long and default ignores will make it longer, and
default Rails-specific ignores will make it longer still. Broke down 
into a delegator. Chose not to create a value type around 
RuboCop::ConfigStore because wrapping it would sorta require a lot of
state juggling and potentially multiple instances b/c of the tricky 
stuff it does when a yaml file is assigned (it also hits disk when it 
does this). Instead, each unit effectively mutates the store object as
we need.