Align our ignore directory scheme to RuboCop's
This was obscenely tricky to get right.
1. Spelunk Rubocop's codebase, learn that (1) if there's a config file
defined, make default excludes relative to that file, (2) if there is 
not, make them relative do the working directory
2. Update all of our tests to match that scheme, made more challenging
by the fact that most of our config tests are designed to simulate a
different working directory without actually changing pwd and that
changing pwd in-process with Dir.cwd doesn't actually alter rubocop's
behavior.

So, ultimately I _hope_ this works, but I think it means we really need
an additional test that calls the gem's actual executable against a 
sample project and from multiple directory locations to be sure we got 
it right.