Add --no-fix option
In the case that you use the config to automatically fix your
violations `fix: true`, this flag will override that setting in the
negative direction, just as --fix overrides the lack of that setting.

I went with `no-` as the prefix because it seems to be used most
commonly for inverting options.

fixes #54