Ignore no-op cop disable/enable comments
It seems a bit noisy to blat an error or warning when someone has a 
no-op comment for a particular line. Especially in non-traditional apps/
libs that might graft code from multiple sources or which might play in
multiple builds. Doubly so because standard doesn't apply a hard 
constraint on the underlying verision of rubocop (yet)

Fixes #71