Create helper module to determine if Offense is autocorretable
The helper makes the following assumptions:

- All offenses' `cop_name` will have the format: `department/cop_name`

Disclaimer:
The (docs for the `cop_name` attr_reader)[https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Offense#cop_name-instance_method]
state that the method "Returns a cop class name without department",
however, when I `binding.pry`'d the first line of
`Standard::Formatter#file_finished` the offenses' cop_name had the
department attached to it, so that's why I moved foward with this approach.

Further understanding if this will always be the case would require diving
into the depths of the Rubocop source code, which I consider as a
non-worthwhile effort as of this moment.

P.S.: The `RuboCop::Cop::Bagde` class could solve this issue if it had a
method that receives a badge and returns the corresponding constant.