Refactor PR, fix a bug
This PR had a bug in that it would only print the --fix flag if the 
FIRST file with an uncorrected offense had at least one safely 
autocorrectable cop in it. Now it'll print the header up to one time as
soon as the first file with such an offense occurs.

I do not love this b/c it means that the warning can show up in the 
middle of the output, but I'm too tired of looking at this to make sure
it goes at the end right now.

Also, the formatter was carrying information around in ivars that it
really did not need to, so I whittled that down as much as I could.