Print file if passed via stdin and autofixed
This printing behavior should match how rubocop works, though if you 
want the same formatter you'll want to do `--format emacs` in your
invocation of `standardrb`.

Also adds coverage of rubocop's runner printing a warning or error, 
which had been dogging me to imitate through the CLI itself, so finally
had a chance to get coverage of that function. We're now at 100% code
coverage of the code we own in standard.

Fixes #80