Enforce spaces inside hashes
This is a case where @samphippen verbally agreed to something, and then
I told @noelrap on a podcast that Standard behaves this way, then months
later someone opened an issue and I took over a month to realize that I
was mistaken and Standard did not actually yet behave this way.

Nobody's perfect, I guess, is the point.

Fixes #106