Add test for ignored_method
Adds a unit test that executes the `ignored_method?` check
https://github.com/testdouble/standard/blob/master/lib/standard/cop/semantic_blocks.rb#L105

The methods that are ignored are configured as part of the cop
SemanticBlocks configuration.
[/config/base.yml](https://github.com/testdouble/standard/blob/f6896ac8d79a5697d73ba1b80f60449698bbed86/config/base.yml#L671)

Test coverage:

Before: 710 / 721 LOC (98.47%) covered.
After: 713 / 723 LOC (98.62%) covered.

(The extra 2 lines here are due to the test files themselves being
included in the coverage report. Whether this is a good thing is a
separate question.)