Temporarily stop advertising LSP server capability to executeCommand
There's probably no problem with the server advertising this capability, but current versions of the VS Code extension will stop working until that's updated to
stop specifying the same command. This requires further investigation, so I'm just disabling the registration on the server side to keep the extension working
for now.

Fixes #544