Print the number of potential fixes (the number of correctable problems) instead of just "some"
Now that we're counting these, we may as well give the user a clue how many would be fixed.

Note that this this changes the order of the output. Previously, the hint to `run standard --fix` was printed immediately before the first problem that was fixable, even if that put it in the middle of the output.

Now, because we're waiting to count them first, the hint will run at the very end of the output. Any scripts depending on this hint appearing earlier (which would be odd) could be impacted. From a UX standpoint it makes more sense to print this at the end, anyway, since it was always weird when it showed up in the middle of the output