Remove incorrect textDocument/diagnostic request handler
The parameters for this request only include a textDocument, which is a
TextDocumentIdentifier, an identifier? and a previousResultId?. The
TextDocumentIdentifer only has a uri property, which is of type
DocumentUri.

According to the LSP spec:

https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_diagnostic

this request does not include the contents of the file (which is what
the handler was expecting).

I have simply made the handling of this request a no-op, since
diagnostics are already sent to the client in textDocument/didChange.

Note: This also fixes formatting problems related to Neovim 0.10
described in #575