Ensure returned uri from code actions contain a scheme
Using Neovim 0.10 with the ruby-lsp plugin, standard returns code
actions to allow for quick fixes. They were returning the uri without
the scheme, and that caused neovim to give an error.

To correct for the error, I added an ensure_uri_scheme method to parse
the uri and add the "file" scheme if the uri did not include one.