Allow RBS inline annotation
Type signature for Ruby (rbs) allow type checkers (e.g.  steep) to
statically check for issues in a code base.  Developpers must sometimes
provide hints when initializing a data structure (e.g. an empty Array)
so that the type checker can properly check its usage.  These hints are
passed with an in-line comment like so:

```
results = [] #: Array[Integer]
```

Rubocop added support for such comments as on option in its
Layout/LeadingEmptyLines check in version 1.67.0.  This gem currently
use a newer version so this option is available but not enabled yet.

Enable this option to make it possible to use both rbs and standard at
the same time.