Delay dependabot updates with cooldown
This gives the updates time to be "shaken out". Most of our updates are
dev dependencies so the delaying security updates isn't an issue. And
the delay will help allow the community to catch issues if there are
bugs (or compromised deps). Not a _huge_ delay so as to risk unpatched
code, but also we don't want updates the day they're released either.

Our weekly cadence + the 15day cooldown makes the effective delay
approximately 3 weeks.

Also, allow rubocop updates to be opened immediately since the purpose
of those PRs is to review the new cops/changes regularly.